backend/requirements.txt
flask
pymysql
requests
python-dotenv
pytz
gunicorn
google-generativeai
